requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.0.0
accelerate>=0.20.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Optional dependencies
# wandb>=0.15.0  # For experiment tracking
# onnx>=1.14.0  # For model export
